Fix: See 'No more movies to load.' when at the end of a movie list
Previously it was showing the 'SHOW MORE' button, even if there were no more movies in the list.
Bug fix was on the BE, but it needed a small update on the FE as well.
There is no 'next' page prop at the end of a movie list (ai: no more pages left), hence the optional chaining operator.